requirements.txt
sentence-transformers 
transformers 
langchain 
langchain-community 
pypdf 
faiss-cpu 
chromadb 
scikit-learn 
numpy 
pandas 
ollama 
matplotlib
spacy
networkx
python-louvain
lxml
openai
gradio